Use the REST API to hit Vertex AI instead of the Node libraries as using the libraries just wouldn't work in a Docker container
Specifically, the Google Cloud auth token would never generate despite being authed into the gcloud cli. This was with using the Google Cloud base image as well as installing it with the Node base image. Even after manually logging in, there would be another error about the Vertex AI client with a 404.